refactor: move shared account utils to module
